new changes added

+++ Licious-clone-main/src/utilities/logout.js
@@ -1,18 +1,38 @@
-import React from "react";
+import React, { useContext } from "react";
+import { MyContext } from "../context/app-context";
 
 const Logout = (props) => {
+  const context = useContext(MyContext);
+
   const clear = () => {
-    sessionStorage.clear();
-    alert("Logout Successfully");
+    const currtUser = JSON.parse(sessionStorage.getItem("user")) ?? {};
+    const userUsername = currtUser.userName;
+    const userCart = currtUser.cart;
+
+    const records = JSON.parse(localStorage.getItem("accounts")) ?? [];
+    const updatedRecords = records.map((e) =>
+      e.userName === userUsername ? { ...e, cart: userCart } : e
+    );
+
+    localStorage.setItem("accounts", JSON.stringify(updatedRecords));
+
+    if (context.logInStatus !== false && context.currentUser !== null) {
+      alert("Logout Successfully");
+      context.setCurrentUser(null);
+      context.setLogInStatus(false);
+      sessionStorage.clear();
+      sessionStorage.removeItem("user");
+      window.location.reload();
+    }
   };
+
   return (
     <div>
-      <dialog id="logout">
-        <div>
-          <button onClick={(e) => clear()}>Logout</button>
-        </div>
-      </dialog>
+      <div onClick={clear} style={{ margin:"0px 10px" }}>
+        Logout
+      </div>
     </div>
   );
 };
+
 export default Logout;
